Appease the machine gods
